✨ feat(examples/with-redux) move to app router
