+++ examples/with-redux/src/app/components/Counter/Counter.tsx
@@ -1,3 +1,5 @@
+'use client'
+
 /* Core */
 import { useState } from 'react'
