+++ examples/with-redux/src/app/components/Counter/index.ts
