+++ examples/with-redux/src/app/components/Nav.tsx
