+++ examples/with-redux/src/app/components/index.ts
