+++ examples/with-redux/src/app/icon.ico
