--- examples/with-redux/src/pages/api/identity-count.ts
@@ -1,14 +1,0 @@
-/* Core */
-import waait from 'waait'
-import type { NextApiHandler } from 'next'
-
-const identityCountHandler: NextApiHandler = async (request, response) => {
-  const { amount = 1 } = request.body
-
-  // simulate IO latency
-  await waait(500)
-
-  response.json({ data: amount })
-}
-
-export default identityCountHandler
